cleanup and reorganize code, optimize SVGs
