+++ docs/ui/components/Home/APIIcons.tsx
@@ -1,0 +1,58 @@
+import { iconSize, theme } from '@expo/styleguide';
+import React from 'react';
+
+import { IconProps } from '~/ui/foundations/icons';
+
+export const APICameraIcon = ({ size = iconSize.regular }: IconProps) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 70 70"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M19.2208 17.5L27.0958 9.625C27.3659 9.34941 27.6881 9.13016 28.0435 8.97997C28.399 8.82977 28.7808 8.7516 29.1666 8.75H40.8333C41.6026 8.76096 42.3365 9.07546 42.875 9.625L50.8083 17.5H58.3333C59.8804 17.5 61.3641 18.1146 62.4581 19.2085C63.5521 20.3025 64.1666 21.7862 64.1666 23.3333V55.4167C64.1666 56.9638 63.5521 58.4475 62.4581 59.5415C61.3641 60.6354 59.8804 61.25 58.3333 61.25H11.6666C10.1196 61.25 8.63582 60.6354 7.54186 59.5415C6.44789 58.4475 5.83331 56.9638 5.83331 55.4167V23.3333C5.83331 20.125 8.45831 17.5 11.6666 17.5H19.2208ZM55.4166 29.1667C56.1902 29.1667 56.9321 28.8594 57.479 28.3124C58.026 27.7654 58.3333 27.0235 58.3333 26.25C58.3333 25.4765 58.026 24.7346 57.479 24.1876C56.9321 23.6406 56.1902 23.3333 55.4166 23.3333C54.6431 23.3333 53.9012 23.6406 53.3543 24.1876C52.8073 24.7346 52.5 25.4765 52.5 26.25C52.5 27.0235 52.8073 27.7654 53.3543 28.3124C53.9012 28.8594 54.6431 29.1667 55.4166 29.1667ZM35 52.5C38.8677 52.5 42.577 50.9635 45.312 48.2286C48.0469 45.4937 49.5833 41.7844 49.5833 37.9167C49.5833 34.0489 48.0469 30.3396 45.312 27.6047C42.577 24.8698 38.8677 23.3333 35 23.3333C31.1322 23.3333 27.4229 24.8698 24.688 27.6047C21.9531 30.3396 20.4166 34.0489 20.4166 37.9167C20.4166 41.7844 21.9531 45.4937 24.688 48.2286C27.4229 50.9635 31.1322 52.5 35 52.5Z"
+      fill={theme.icon.secondary}
+    />
+    <path
+      d="M35 46.6666C37.3206 46.6666 39.5462 45.7448 41.1872 44.1038C42.8281 42.4629 43.75 40.2373 43.75 37.9166C43.75 35.596 42.8281 33.3704 41.1872 31.7294C39.5462 30.0885 37.3206 29.1666 35 29.1666C32.6794 29.1666 30.4538 30.0885 28.8128 31.7294C27.1719 33.3704 26.25 35.596 26.25 37.9166C26.25 40.2373 27.1719 42.4629 28.8128 44.1038C30.4538 45.7448 32.6794 46.6666 35 46.6666Z"
+      fill={theme.icon.default}
+    />
+  </svg>
+);
+
+export const APINotificationsIcon = ({ size = iconSize.regular }: IconProps) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 70 70"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M35 64.1666C39.8325 64.1666 43.75 60.2491 43.75 55.4166C43.75 50.5841 39.8325 46.6666 35 46.6666C30.1675 46.6666 26.25 50.5841 26.25 55.4166C26.25 60.2491 30.1675 64.1666 35 64.1666Z"
+      fill={theme.icon.default}
+    />
+    <path
+      d="M29.225 12.4835L29.1667 11.6668C29.1658 10.8658 29.33 10.0731 29.6489 9.33827C29.9678 8.60343 30.4346 7.94215 31.0203 7.39565C31.606 6.84914 32.298 6.42912 33.0531 6.16176C33.8082 5.8944 34.6103 5.78544 35.4094 5.84165C36.2085 5.89787 36.9874 6.11807 37.6976 6.48852C38.4079 6.85897 39.0342 7.37173 39.5376 7.99486C40.041 8.61798 40.4107 9.33811 40.6236 10.1104C40.8365 10.8826 40.888 11.6904 40.775 12.4835C45.005 13.7309 48.7174 16.3144 51.3567 19.8475C53.996 23.3807 55.4204 27.6734 55.4167 32.0835V46.6668C55.4167 47.4403 55.724 48.1822 56.2709 48.7292C56.8179 49.2762 57.5598 49.5835 58.3333 49.5835C59.1069 49.5835 59.8488 49.8908 60.3957 50.4377C60.9427 50.9847 61.25 51.7266 61.25 52.5001C61.25 53.2737 60.9427 54.0155 60.3957 54.5625C59.8488 55.1095 59.1069 55.4168 58.3333 55.4168H11.6667C10.8931 55.4168 10.1513 55.1095 9.60427 54.5625C9.05729 54.0155 8.75 53.2737 8.75 52.5001C8.75 51.7266 9.05729 50.9847 9.60427 50.4377C10.1513 49.8908 10.8931 49.5835 11.6667 49.5835C12.4402 49.5835 13.1821 49.2762 13.7291 48.7292C14.276 48.1822 14.5833 47.4403 14.5833 46.6668V32.0835C14.5796 27.6734 16.004 23.3807 18.6433 19.8475C21.2826 16.3144 24.995 13.7309 29.225 12.4835Z"
+      fill={theme.icon.secondary}
+    />
+  </svg>
+);
+
+export const APIListIcon = ({ size = iconSize.regular }: IconProps) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 70 70"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M7.58337 17.7623L33.8334 6.09567C34.2014 5.93507 34.5985 5.85217 35 5.85217C35.4015 5.85217 35.7987 5.93507 36.1667 6.09567L62.4167 17.7623C62.9259 17.9943 63.3577 18.3679 63.6604 18.8385C63.9631 19.3092 64.1241 19.8569 64.1241 20.4165C64.1241 20.9761 63.9631 21.5238 63.6604 21.9945C63.3577 22.4651 62.9259 22.8387 62.4167 23.0707L36.1667 34.7373C35.7987 34.8979 35.4015 34.9808 35 34.9808C34.5985 34.9808 34.2014 34.8979 33.8334 34.7373L7.58337 23.0707C7.07415 22.8387 6.64243 22.4651 6.33969 21.9945C6.03695 21.5238 5.87598 20.9761 5.87598 20.4165C5.87598 19.8569 6.03695 19.3092 6.33969 18.8385C6.64243 18.3679 7.07415 17.9943 7.58337 17.7623Z"
+      fill={theme.icon.secondary}
+    />
+    <path
+      d="M11.4043 30.625L33.8043 40.5708C34.1803 40.7399 34.5879 40.8273 35.0002 40.8273C35.4124 40.8273 35.82 40.7399 36.196 40.5708L58.596 30.625L62.446 32.3458C62.9552 32.5778 63.3869 32.9514 63.6897 33.422C63.9924 33.8927 64.1534 34.4404 64.1534 35C64.1534 35.5596 63.9924 36.1073 63.6897 36.578C63.3869 37.0486 62.9552 37.4222 62.446 37.6542L36.196 49.3208C35.82 49.4899 35.4124 49.5773 35.0002 49.5773C34.5879 49.5773 34.1803 49.4899 33.8043 49.3208L7.55432 37.6542C7.0451 37.4222 6.61337 37.0486 6.31063 36.578C6.00789 36.1073 5.84692 35.5596 5.84692 35C5.84692 34.4404 6.00789 33.8927 6.31063 33.422C6.61337 32.9514 7.0451 32.5778 7.55432 32.3458L11.4043 30.625ZM11.4043 45.2083L33.8043 55.1542C34.1803 55.3232 34.5879 55.4106 35.0002 55.4106C35.4124 55.4106 35.82 55.3232 36.196 55.1542L58.596 45.2083L62.446 46.9292C62.9552 47.1612 63.3869 47.5348 63.6897 48.0054C63.9924 48.476 64.1534 49.0238 64.1534 49.5833C64.1534 50.1429 63.9924 50.6907 63.6897 51.1613C63.3869 51.6319 62.9552 52.0055 62.446 52.2375L36.196 63.9042C35.82 64.0732 35.4124 64.1606 35.0002 64.1606C34.5879 64.1606 34.1803 64.0732 33.8043 63.9042L7.55432 52.2375C7.0451 52.0055 6.61337 51.6319 6.31063 51.1613C6.00789 50.6907 5.84692 50.1429 5.84692 49.5833C5.84692 49.0238 6.00789 48.476 6.31063 48.0054C6.61337 47.5348 7.0451 47.1612 7.55432 46.9292L11.4043 45.2083Z"
+      fill={theme.icon.default}
+    />
+  </svg>
+);
