+++ docs/ui/components/Home/DevicesImage.tsx
@@ -12,7 +12,7 @@ export const DevicesImage = () => (
     viewBox="0 0 354 164"
     fill="none"
     xmlns="http://www.w3.org/2000/svg">
-    <g id="6a0631cecbb09de94722a89243995162" filter="url(#3fe977fb0acabded0c62aa0c9c945938)">
+    <g filter="url(#3fe977fb0acabded0c62aa0c9c945938)">
       <g clipPath="url(#clip0_39:1396)">
         <rect
           x="214.901"
@@ -23,24 +23,22 @@ export const DevicesImage = () => (
           transform="rotate(13 214.901 22.3787)"
           fill="#464D55"
         />
-        <g id="74ad6257659e362c187cca53806b2a2f">
-          <rect
-            x="211.247"
-            y="38.208"
-            width="195"
-            height="171.754"
-            transform="rotate(13 211.247 38.208)"
-            fill="#745BFF"
-          />
-          <rect
-            x="211.247"
-            y="38.208"
-            width="195"
-            height="171.754"
-            transform="rotate(13 211.247 38.208)"
-            fill="url(#efeb358a5390d90e5e3ede15cf021e74)"
-          />
-        </g>
+        <rect
+          x="211.247"
+          y="38.208"
+          width="195"
+          height="171.754"
+          transform="rotate(13 211.247 38.208)"
+          fill="#745BFF"
+        />
+        <rect
+          x="211.247"
+          y="38.208"
+          width="195"
+          height="171.754"
+          transform="rotate(13 211.247 38.208)"
+          fill="url(#efeb358a5390d90e5e3ede15cf021e74)"
+        />
         <rect
           x="251.549"
           y="60.3827"
@@ -117,7 +115,7 @@ export const DevicesImage = () => (
           transform="rotate(13 260.869 75.1075)"
           fill="#FFE1A7"
         />
-        <g id="305036a5fc80a86b4e15e7bd8f2da842" clipPath="url(#clip1_39:1396)">
+        <g clipPath="url(#clip1_39:1396)">
           <path
             fillRule="evenodd"
             clipRule="evenodd"
@@ -154,95 +152,93 @@ export const DevicesImage = () => (
         />
       </g>
     </g>
-    <g id="509ce3782e3cf8366cd010632bcd8d7b" filter="url(#1bc05e09f77f6a149af69cdd548e2519)">
-      <g id="882e2a6b4ca338fa93c5c03cd916b56f">
-        <g id="8cd14a4c24a00e9f93c8cf498ff0b5bd" filter="url(#5f24a6d69757ffdb29b739d4fc5a5af1)">
-          <rect
-            x="13.8011"
-            y="45.1538"
-            width="88.9538"
-            height="186.813"
-            rx="9.95547"
-            transform="rotate(-13 13.8011 45.1538)"
-            fill="url(#798da3dad1527946ee33aecf8f9ec234)"
-          />
-          <rect
-            x="18.0653"
-            y="47.8184"
-            width="81.8428"
-            height="179.702"
-            rx="6.39994"
-            transform="rotate(-13 18.0653 47.8184)"
-            stroke="#464D55"
-            strokeWidth="7.11105"
-          />
-        </g>
+    <g filter="url(#1bc05e09f77f6a149af69cdd548e2519)">
+      <g filter="url(#5f24a6d69757ffdb29b739d4fc5a5af1)">
         <rect
-          x="27.9378"
-          y="53.9677"
-          width="65.4637"
-          height="41.3883"
-          rx="3.31849"
-          transform="rotate(-13 27.9378 53.9677)"
-          fill="white"
+          x="13.8011"
+          y="45.1538"
+          width="88.9538"
+          height="186.813"
+          rx="9.95547"
+          transform="rotate(-13 13.8011 45.1538)"
+          fill="url(#798da3dad1527946ee33aecf8f9ec234)"
         />
         <rect
-          x="43.8268"
-          y="122.618"
-          width="65.4637"
-          height="19.1339"
-          rx="3.31849"
-          transform="rotate(-13 43.8268 122.618)"
-          fill="white"
+          x="18.0653"
+          y="47.8184"
+          width="81.8428"
+          height="179.702"
+          rx="6.39994"
+          transform="rotate(-13 18.0653 47.8184)"
+          stroke="#464D55"
+          strokeWidth="7.11105"
         />
+      </g>
+      <rect
+        x="27.9378"
+        y="53.9677"
+        width="65.4637"
+        height="41.3883"
+        rx="3.31849"
+        transform="rotate(-13 27.9378 53.9677)"
+        fill="white"
+      />
+      <rect
+        x="43.8268"
+        y="122.618"
+        width="65.4637"
+        height="19.1339"
+        rx="3.31849"
+        transform="rotate(-13 43.8268 122.618)"
+        fill="white"
+      />
+      <rect
+        opacity="0.8"
+        x="38.4219"
+        y="99.2343"
+        width="65.4637"
+        height="6.92818"
+        rx="3.31849"
+        transform="rotate(-13 38.4219 99.2343)"
+        fill="white"
+      />
+      <rect
+        opacity="0.8"
+        x="47.0919"
+        y="106.469"
+        width="51.8308"
+        height="6.92818"
+        rx="3.31849"
+        transform="rotate(-13 47.0919 106.469)"
+        fill="white"
+      />
+      <mask
+        id="c91e2833bce82b5f6ccc942f780f85d1"
+        style={{ maskType: 'alpha' }}
+        maskUnits="userSpaceOnUse"
+        x="28"
+        y="40"
+        width="73"
+        height="54">
         <rect
-          opacity="0.8"
-          x="38.4219"
-          y="99.2343"
-          width="65.4637"
-          height="6.92818"
+          x="27.7824"
+          y="54.4393"
+          width="65.7216"
+          height="40.9636"
           rx="3.31849"
-          transform="rotate(-13 38.4219 99.2343)"
+          transform="rotate(-13 27.7824 54.4393)"
           fill="white"
         />
-        <rect
-          opacity="0.8"
-          x="47.0919"
-          y="106.469"
-          width="51.8308"
-          height="6.92818"
-          rx="3.31849"
-          transform="rotate(-13 47.0919 106.469)"
-          fill="white"
+      </mask>
+      <g mask="url(#c91e2833bce82b5f6ccc942f780f85d1)">
+        <path
+          d="M62.6906 61.9562L45.6377 92.3956L105.302 78.6211L100.386 57.3308L89.5646 48.6166L78.1975 73.1563L62.6906 61.9562Z"
+          fill="#C99CFF"
         />
-        <mask
-          id="c91e2833bce82b5f6ccc942f780f85d1"
-          style={{ maskType: 'alpha' }}
-          maskUnits="userSpaceOnUse"
-          x="28"
-          y="40"
-          width="73"
-          height="54">
-          <rect
-            x="27.7824"
-            y="54.4393"
-            width="65.7216"
-            height="40.9636"
-            rx="3.31849"
-            transform="rotate(-13 27.7824 54.4393)"
-            fill="white"
-          />
-        </mask>
-        <g mask="url(#c91e2833bce82b5f6ccc942f780f85d1)">
-          <path
-            d="M62.6906 61.9562L45.6377 92.3956L105.302 78.6211L100.386 57.3308L89.5646 48.6166L78.1975 73.1563L62.6906 61.9562Z"
-            fill="#C99CFF"
-          />
-          <path
-            d="M52.7926 70.6425L40.1633 93.6194L84.598 83.3609L97.6809 80.3404L72.7879 60.6251L64.3809 79.1551L52.7926 70.6425Z"
-            fill="#A771E7"
-          />
-        </g>
+        <path
+          d="M52.7926 70.6425L40.1633 93.6194L84.598 83.3609L97.6809 80.3404L72.7879 60.6251L64.3809 79.1551L52.7926 70.6425Z"
+          fill="#A771E7"
+        />
       </g>
       <circle
         cx="42.7692"
@@ -251,16 +247,14 @@ export const DevicesImage = () => (
         transform="rotate(-13 42.7692 63.1166)"
         fill="#FFE1A7"
       />
-      <g id="82cda98e5d33ee683d0313b370bf93d4" clipPath="url(#clip2_39:1396)">
+      <g clipPath="url(#clip2_39:1396)">
         <path
-          id="feaf4b96123edb786ee9c88c922b1300"
           fillRule="evenodd"
           clipRule="evenodd"
           d="M84.1094 117.748C84.0501 117.491 83.7938 117.331 83.537 117.391C83.2802 117.45 83.12 117.706 83.1793 117.963L83.4656 119.203L82.2258 119.489C81.969 119.549 81.8088 119.805 81.8681 120.062C81.9274 120.319 82.1837 120.479 82.4405 120.419L83.6804 120.133L83.9667 121.373C84.026 121.63 84.2822 121.79 84.5391 121.731C84.7959 121.672 84.9561 121.415 84.8968 121.159L84.6105 119.918L85.8509 119.632C86.1078 119.573 86.2679 119.317 86.2086 119.06C86.1493 118.803 85.893 118.643 85.6362 118.702L84.3957 118.988L84.1094 117.748Z"
           fill="#596068"
         />
         <path
-          id="dd04fd9e7e6888e0183eea7a553566ae"
           fillRule="evenodd"
           clipRule="evenodd"
           d="M76.5483 122.056L76.4534 122.945L75.4045 123.187C74.7196 123.345 74.2926 124.028 74.4507 124.713L75.5959 129.674C75.7541 130.359 76.4375 130.786 77.1224 130.628L83.9431 129.053C84.628 128.895 85.0551 128.211 84.897 127.527L83.7517 122.566C83.5936 121.881 82.9102 121.454 82.2253 121.612L81.1756 121.854L80.7007 121.097C80.5566 120.867 80.2827 120.754 80.0184 120.815L77.0379 121.503C76.7737 121.564 76.5771 121.786 76.5483 122.056ZM80.0313 127.67C81.0587 127.433 81.6993 126.408 81.4621 125.38C81.2249 124.353 80.1998 123.712 79.1724 123.95C78.1451 124.187 77.5045 125.212 77.7417 126.239C77.9789 127.267 79.004 127.907 80.0313 127.67Z"
@@ -268,100 +262,75 @@ export const DevicesImage = () => (
         />
       </g>
     </g>
-    <g id="bc8ae3631d7883c6e7415ca4c01acb6a" filter="url(#71eac5b5383aeb9f82fcc3ed96d96c4b)">
-      <g id="ae8cb7de269bc23b780b50faa08fae02">
-        <g id="94ca9105e866504c074c73d20673af55" filter="url(#312970ea83b71b83cdd5aadda6bb6908)">
-          <rect
-            x="113.731"
-            y="13.4504"
-            width="88.9538"
-            height="186.813"
-            rx="9.95547"
-            fill="url(#e2839d57a5b978f1058be04fb2a500dd)"
-          />
-          <rect
-            x="117.287"
-            y="17.006"
-            width="81.8428"
-            height="179.702"
-            rx="6.39994"
-            stroke="#464D55"
-            strokeWidth="7.11105"
-          />
-        </g>
-        <rect x="125.523" y="33.6431" width="65.4637" height="41.3883" rx="3.31849" fill="white" />
-        <rect x="125.562" y="104.108" width="65.4637" height="19.1339" rx="3.31849" fill="white" />
+    <g filter="url(#71eac5b5383aeb9f82fcc3ed96d96c4b)">
+      <g filter="url(#312970ea83b71b83cdd5aadda6bb6908)">
         <rect
-          opacity="0.8"
-          x="125.556"
-          y="80.1079"
-          width="65.4637"
-          height="6.92818"
-          rx="3.31849"
-          fill="white"
+          x="113.731"
+          y="13.4504"
+          width="88.9538"
+          height="186.813"
+          rx="9.95547"
+          fill="url(#e2839d57a5b978f1058be04fb2a500dd)"
         />
         <rect
-          id="703c6f2f19d585d6b9a95cd164f54720"
-          opacity="0.8"
-          x="132.376"
-          y="89.1079"
-          width="51.8308"
-          height="6.92818"
-          rx="3.31849"
-          fill="white"
+          x="117.287"
+          y="17.006"
+          width="81.8428"
+          height="179.702"
+          rx="6.39994"
+          stroke="#464D55"
+          strokeWidth="7.11105"
         />
-        <g id="6f03fbf470879a844d62971285ec93d3">
-          <mask
-            id="cac8353d923d12b52a34815740143c7d"
-            style={{ maskType: 'alpha' }}
-            maskUnits="userSpaceOnUse"
-            x="125"
-            y="34"
-            width="66"
-            height="42">
-            <rect
-              id="371bfb89217320b653a5ee4616be6fd7"
-              x="125.266"
-              y="34.0677"
-              width="65.7216"
-              height="40.9636"
-              rx="3.31849"
-              fill="white"
-            />
-          </mask>
-          <g mask="url(#cac8353d923d12b52a34815740143c7d)">
-            <path
-              id="2004acccdea4733d0deb6edbf055641f"
-              d="M157.588 49.2446L134.125 75.0678H195.358V53.2174L186.774 42.2922L170.178 63.646L157.588 49.2446Z"
-              fill="#C99CFF"
-            />
-            <path
-              id="f1058eb09cae009fb0faab26b11ee039"
-              d="M145.99 55.4817L128.516 75.0287H174.119H187.546L167.726 50.219L155.366 66.3829L145.99 55.4817Z"
-              fill="#A771E7"
-            />
-          </g>
-        </g>
       </g>
-      <circle
-        id="0b242f322b78ee6808bf00f7ab55bf14"
-        cx="137.916"
-        cy="45.894"
-        r="7.01414"
-        fill="#FFE1A7"
+      <rect x="125.523" y="33.6431" width="65.4637" height="41.3883" rx="3.31849" fill="white" />
+      <rect x="125.562" y="104.108" width="65.4637" height="19.1339" rx="3.31849" fill="white" />
+      <rect
+        opacity="0.8"
+        x="125.556"
+        y="80.1079"
+        width="65.4637"
+        height="6.92818"
+        rx="3.31849"
+        fill="white"
       />
-      <g id="1d56cc72ab577bd7ab8654f6784dccd5" clipPath="url(#clip3_39:1396)">
-        <g id="31cc7f4835381ac3d5d3fd7e6e5d37a0">
-          <path
-            id="a08be458bb7b06f625d9dd7c6641a408"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M165.908 108.425C165.908 108.161 165.694 107.948 165.43 107.948C165.167 107.948 164.953 108.161 164.953 108.425V109.698H163.68C163.417 109.698 163.203 109.911 163.203 110.175C163.203 110.438 163.417 110.652 163.68 110.652H164.953V111.925C164.953 112.188 165.167 112.402 165.43 112.402C165.694 112.402 165.908 112.188 165.908 111.925V110.652H167.18C167.444 110.652 167.658 110.438 167.658 110.175C167.658 109.911 167.444 109.698 167.18 109.698H165.908V108.425Z"
-            fill="#596068"
-          />
-        </g>
+      <rect
+        opacity="0.8"
+        x="132.376"
+        y="89.1079"
+        width="51.8308"
+        height="6.92818"
+        rx="3.31849"
+        fill="white"
+      />
+      <mask
+        id="cac8353d923d12b52a34815740143c7d"
+        style={{ maskType: 'alpha' }}
+        maskUnits="userSpaceOnUse"
+        x="125"
+        y="34"
+        width="66"
+        height="42">
+        <rect x="125.266" y="34.0677" width="65.7216" height="40.9636" rx="3.31849" fill="white" />
+      </mask>
+      <g mask="url(#cac8353d923d12b52a34815740143c7d)">
         <path
-          id="fcc72da3ab86bbb30347f4754f9ed575"
+          d="M157.588 49.2446L134.125 75.0678H195.358V53.2174L186.774 42.2922L170.178 63.646L157.588 49.2446Z"
+          fill="#C99CFF"
+        />
+        <path
+          d="M145.99 55.4817L128.516 75.0287H174.119H187.546L167.726 50.219L155.366 66.3829L145.99 55.4817Z"
+          fill="#A771E7"
+        />
+      </g>
+      <circle cx="137.916" cy="45.894" r="7.01414" fill="#FFE1A7" />
+      <g clipPath="url(#clip3_39:1396)">
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M165.908 108.425C165.908 108.161 165.694 107.948 165.43 107.948C165.167 107.948 164.953 108.161 164.953 108.425V109.698H163.68C163.417 109.698 163.203 109.911 163.203 110.175C163.203 110.438 163.417 110.652 163.68 110.652H164.953V111.925C164.953 112.188 165.167 112.402 165.43 112.402C165.694 112.402 165.908 112.188 165.908 111.925V110.652H167.18C167.444 110.652 167.658 110.438 167.658 110.175C167.658 109.911 167.444 109.698 167.18 109.698H165.908V108.425Z"
+          fill="#596068"
+        />
+        <path
           fillRule="evenodd"
           clipRule="evenodd"
           d="M157.572 110.921L157.279 111.766H156.202C155.5 111.766 154.93 112.336 154.93 113.039V118.13C154.93 118.833 155.5 119.402 156.202 119.402H163.203C163.906 119.402 164.475 118.833 164.475 118.13V113.039C164.475 112.336 163.906 111.766 163.203 111.766H162.126L161.833 110.921C161.745 110.665 161.503 110.493 161.232 110.493H158.173C157.902 110.493 157.66 110.665 157.572 110.921ZM159.703 117.175C160.757 117.175 161.612 116.32 161.612 115.266C161.612 114.212 160.757 113.357 159.703 113.357C158.648 113.357 157.793 114.212 157.793 115.266C157.793 116.32 158.648 117.175 159.703 117.175Z"
@@ -369,7 +338,6 @@ export const DevicesImage = () => (
         />
       </g>
       <path
-        id="2c3d1f35940428e1bb443d5d8bc4d803"
         d="M134.302 13.9601C134.856 13.6266 135.491 13.4504 136.139 13.4504H180.206C180.899 13.4504 181.576 13.6521 182.156 14.0309L185.142 15.9811C187.227 17.343 186.263 20.5831 183.773 20.5831C182.389 20.5831 181.268 21.7046 181.268 23.088V24.1495C181.268 26.1191 179.671 27.7158 177.701 27.7158H138.716C136.746 27.7158 135.149 26.1191 135.149 24.1495V23.1602C135.149 21.7369 133.996 20.5831 132.572 20.5831C129.96 20.5831 129.005 17.1432 131.245 15.7973L134.302 13.9601Z"
         fill="#464D55"
       />
