+++ docs/ui/components/Home/WhyImage.tsx
@@ -13,35 +13,25 @@ export const WhyImage = () => (
     fill="none"
     xmlns="http://www.w3.org/2000/svg">
     <path
-      id="c95791cc0ca8136d4c257abc84440b61"
       d="M72.7084 38.4707C72.7084 57.7122 57.1101 73.3104 37.8687 73.3104C18.6272 73.3104 3.02893 57.7122 3.02893 38.4707C3.02893 19.2293 18.6272 3.63098 37.8687 3.63098C57.1101 3.63098 72.7084 19.2293 72.7084 38.4707Z"
       fill="#DCFFE4"
       stroke="#85E89D"
       strokeWidth="6"
     />
     <path
-      id="1846c8357fd69b2d2bad377fc01f1e96"
       d="M131.699 38.4707C131.699 57.7122 116.101 73.3104 96.8597 73.3104C77.6183 73.3104 62.02 57.7122 62.02 38.4707C62.02 19.2293 77.6183 3.63098 96.8597 3.63098C116.101 3.63098 131.699 19.2293 131.699 38.4707Z"
       fill="#BEF5CB"
       stroke="#85E89D"
       strokeWidth="6"
     />
+    <path d="M96.8596 25.7819V51.1596" stroke="#28A745" strokeWidth="6.67" strokeLinecap="round" />
     <path
-      id="6a42df9e9f690d9c9a055a3f648f430f"
-      d="M96.8596 25.7819V51.1596"
-      stroke="#28A745"
-      strokeWidth="6.67"
-      strokeLinecap="round"
-    />
-    <path
-      id="487c3d8c60f6b8c7b1df681cdede7de8"
       d="M84.1716 38.4708L109.549 38.4708"
       stroke="#28A745"
       strokeWidth="6.67"
       strokeLinecap="round"
     />
     <path
-      id="6eeafa7a395a6e3fbab94af09e2d16ed"
       d="M25.1804 38.4708L50.5581 38.4708"
       stroke="#28A745"
       strokeWidth="6.67"
